The Indexer Runs!
It occurred to me that I should submit the indexer I've built in it's primitive but functional form.  This is made to run only on my own node in a windows environment with video services and other servers, benchmark testing, gaming related applications, and other varied load scenarios a lightweight indexer is likely to face when mixed with future systems.  Not sure about judging rules so I wanted to submit this which is about 3 days into the indexer searching the bitcoin blockchain for Bitlive! channels.  This version takes about a half hour to run and searches from the first bitlive to the current end of chain, this system will be developed into the "new block interpreter" and then a supplemental blockchain parser will search the "old blocks" in the future to adapt this indexer to more use cases.  You can modify this code quickly with chat gpt and change the search terms to whatever you want, eventually the gui will allow for setting the datadir and picking your block range to search and search terms.  In it's current form it can index the bitcoin blockchain in 1 week (estimate).  Good luck and contact em for a newer version if needed. -Jason