Add exe for python unenjoyors.
This is the windows exe for running the python indexer as an exe.  You must have bitcoin core running with a full node but you do not need ord to detect ordinals etc.  You must input the location of your cli and of your -datadir= so the auth cookie and blockchain data directory are correct.  If you cannot copy paste in the gui try using ctrl-p ctrl-c instead.